+++ braindecode/models/shallow_fbcsp.py
@@ -12,7 +12,11 @@ from braindecode.models.modules import CombinedConv, Ensure4d, Expression, SafeL
 
 
 class ShallowFBCSPNet(EEGModuleMixin, nn.Sequential):
-    """Shallow ConvNet model from Schirrmeister et al. 2017 [Schirrmeister2017]_.
+    """Shallow ConvNet model from Schirrmeister et al (2017) [Schirrmeister2017]_.
+
+    .. figure:: https://onlinelibrary.wiley.com/cms/asset/221ea375-6701-40d3-ab3f-e411aad62d9e/hbm23730-fig-0002-m.jpg
+        :align: center
+        :alt: ShallowNet Architecture
 
     Model described in [Schirrmeister2017]_.